feat: use upgradeable storage

+++ contracts/spdBTC/interfaces/IspdBTC.sol
@@ -8,7 +8,7 @@ struct ProductParams {
     address asset;       // Address of the WBTC ERC20 contract.
     string name;         // ERC20 token name.
     string symbol;       // ERC20 token symbol.
-    uint minDeposit;     // Minimum deposit amount.
-    uint maxDeposit;     // Maximum deposit amount.
+    uint256 minDeposit;  // Minimum deposit amount.
+    uint256 maxDeposit;  // Maximum deposit amount.
     address custodian;   // Custodian address.
 }
